[update] helper module to folder
